+++ core_compat/src/entity/sprite.rs
@@ -6,10 +6,10 @@ use entity::sprite_type::SpriteType;
 pub struct Sprite {
     pub class: SpriteType,
     pub rle_entry: Entry,
-    pub x_dim: usize,
-    pub y_dim: usize,
-    pub x_off: usize,
-    pub y_off: usize,
+    pub x_dim: i32,
+    pub y_dim: i32,
+    pub x_off: i32,
+    pub y_off: i32,
     // pub image_raw: Vec<Pixel>,
     pub image_raw: Vec<u8>, // defined to be in RGB_565_norm format
 }
